Move EntranceEntry
