+++ src/openrct2/libopenrct2.vcxproj
@@ -297,6 +297,7 @@
     <ClInclude Include="object\BannerObject.h" />
     <ClInclude Include="object\BannerSceneryEntry.h" />
     <ClInclude Include="object\DefaultObjects.h" />
+    <ClInclude Include="object\EntranceEntry.h" />
     <ClInclude Include="object\EntranceObject.h" />
     <ClInclude Include="object\FootpathEntry.h" />
     <ClInclude Include="object\FootpathItemEntry.h" />
