+++ src/openrct2/object/EntranceObject.h
@@ -11,6 +11,7 @@
 
 #include "../world/Entrance.h"
 #include "../world/Location.hpp"
+#include "EntranceEntry.h"
 #include "Object.h"
 
 class EntranceObject final : public Object
